need to change forks system

+++ srcs/utils.c
@@ -6,7 +6,7 @@
 /*   By: vbaron <vbaron@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/24 12:05:38 by vbaron            #+#    #+#             */
-/*   Updated: 2021/08/24 15:17:26 by vbaron           ###   ########.fr       */
+/*   Updated: 2021/08/24 16:00:36 by vbaron           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 long	ft_atoi(const char *nptr)
 {
 	int				i;
-	long long		nbr;
+	long		nbr;
 	int				sign;
 
 	i = 0;
